--- a/RecoTracker/ConversionSeedGenerators/python/ConversionStep_cff.py
+++ b/RecoTracker/ConversionSeedGenerators/python/ConversionStep_cff.py
@@ -33,92 +33,95 @@ trackingPhase2PU140.toReplaceWith(convClusters, _phase2trackClusterRemover.clone
     )
 )
 
-convLayerPairs = cms.EDProducer("SeedingLayersEDProducer",
-                                layerList = cms.vstring('BPix1+BPix2', 
+_convLayerPairsStripOnlyLayers = ['TIB1+TID1_pos', 
+                                 'TIB1+TID1_neg', 
+                                 'TIB1+TID2_pos', 
+                                 'TIB1+TID2_neg',
+                                 'TIB1+TIB2',
+                                 'TIB1+TIB3',
+                                 
+                                 'TIB2+TID1_pos', 
+                                 'TIB2+TID1_neg', 
+                                 'TIB2+TID2_pos', 
+                                 'TIB2+TID2_neg', 
+                                 'TIB2+TIB3',
+                                 'TIB2+TIB4', 
+                                 
+                                 'TIB3+TIB4', 
+                                 'TIB3+TOB1', 
+                                 'TIB3+TID1_pos', 
+                                 'TIB3+TID1_neg', 
+                                 
+                                 'TIB4+TOB1',
+                                 'TIB4+TOB2',
+                                 
+                                 'TOB1+TOB2', 
+                                 'TOB1+TOB3', 
+                                 'TOB1+TEC1_pos', 
+                                 'TOB1+TEC1_neg', 
 
-                                                        'BPix2+BPix3', 
-                                                        'BPix2+FPix1_pos', 
-                                                        'BPix2+FPix1_neg', 
-                                                        'BPix2+FPix2_pos', 
-                                                        'BPix2+FPix2_neg', 
-
-                                                        'FPix1_pos+FPix2_pos', 
-                                                        'FPix1_neg+FPix2_neg',
-
-                                                        'BPix3+TIB1', 
-                                                        'BPix3+TIB2',
+                                 'TOB2+TOB3',  
+                                 'TOB2+TOB4',
+                                 'TOB2+TEC1_pos', 
+                                 'TOB2+TEC1_neg', 
+                                 
+                                 #NB: re-introduce these combinations when large displaced
+                                 #    tracks, reconstructed only in TOB will be available
+                                 #    For instance think at the OutIn Ecal Seeded tracks
+                                 #'TOB3+TOB4', 
+                                 #'TOB3+TOB5',
+                                 #'TOB3+TEC1_pos', 
+                                 #'TOB3+TEC1_neg', 
+                                 #
+                                 #'TOB4+TOB5',
+                                 #'TOB4+TOB6',
+                                 #
+                                 #'TOB5+TOB6',
+                                 
+                                 'TID1_pos+TID2_pos', 
+                                 'TID2_pos+TID3_pos', 
+                                 'TID3_pos+TEC1_pos', 
+                                 
+                                 'TID1_neg+TID2_neg', 
+                                 'TID2_neg+TID3_neg', 
+                                 'TID3_neg+TEC1_neg', 
+                                 
+                                 'TEC1_pos+TEC2_pos', 
+                                 'TEC2_pos+TEC3_pos', 
+                                 'TEC3_pos+TEC4_pos',
+                                 'TEC4_pos+TEC5_pos',
+                                 'TEC5_pos+TEC6_pos',
+                                 'TEC6_pos+TEC7_pos',
+                                 'TEC7_pos+TEC8_pos',
+                                 
+                                 'TEC1_neg+TEC2_neg', 
+                                 'TEC2_neg+TEC3_neg', 
+                                 'TEC3_neg+TEC4_neg',
+                                 'TEC4_neg+TEC5_neg',
+                                 'TEC5_neg+TEC6_neg',
+                                 'TEC6_neg+TEC7_neg',
+                                 'TEC7_neg+TEC8_neg']
+                    
+_convLayerPairsLayerList = _convLayerPairsStripOnlyLayers
+_convLayerPairsLayerList =['BPix1+BPix2', 
+                           
+                           'BPix2+BPix3', 
+                           'BPix2+FPix1_pos', 
+                           'BPix2+FPix1_neg', 
+                           'BPix2+FPix2_pos', 
+                           'BPix2+FPix2_neg', 
+                           
+                           'FPix1_pos+FPix2_pos', 
+                           'FPix1_neg+FPix2_neg',
+                                
+                           'BPix3+TIB1', 
+                           'BPix3+TIB2']
+_convLayerPairsLayerList.extend(_convLayerPairsStripOnlyLayers)
                                                         
-                                                        'TIB1+TID1_pos', 
-                                                        'TIB1+TID1_neg', 
-                                                        'TIB1+TID2_pos', 
-                                                        'TIB1+TID2_neg',
-                                                        'TIB1+TIB2',
-                                                        'TIB1+TIB3',
-
-                                                        'TIB2+TID1_pos', 
-                                                        'TIB2+TID1_neg', 
-                                                        'TIB2+TID2_pos', 
-                                                        'TIB2+TID2_neg', 
-                                                        'TIB2+TIB3',
-                                                        'TIB2+TIB4', 
-
-                                                        'TIB3+TIB4', 
-                                                        'TIB3+TOB1', 
-                                                        'TIB3+TID1_pos', 
-                                                        'TIB3+TID1_neg', 
-
-                                                        'TIB4+TOB1',
-                                                        'TIB4+TOB2',
-
-                                                        'TOB1+TOB2', 
-                                                        'TOB1+TOB3', 
-                                                        'TOB1+TEC1_pos', 
-                                                        'TOB1+TEC1_neg', 
-
-                                                        'TOB2+TOB3',  
-                                                        'TOB2+TOB4',
-                                                        'TOB2+TEC1_pos', 
-                                                        'TOB2+TEC1_neg', 
-
-                                                        #NB: re-introduce these combinations when large displaced
-                                                        #    tracks, reconstructed only in TOB will be available
-                                                        #    For instance think at the OutIn Ecal Seeded tracks
-                                                        #'TOB3+TOB4', 
-                                                        #'TOB3+TOB5',
-                                                        #'TOB3+TEC1_pos', 
-                                                        #'TOB3+TEC1_neg', 
-                                                        #
-                                                        #'TOB4+TOB5',
-                                                        #'TOB4+TOB6',
-                                                        #
-                                                        #'TOB5+TOB6',
-
-                                                        'TID1_pos+TID2_pos', 
-                                                        'TID2_pos+TID3_pos', 
-                                                        'TID3_pos+TEC1_pos', 
-
-                                                        'TID1_neg+TID2_neg', 
-                                                        'TID2_neg+TID3_neg', 
-                                                        'TID3_neg+TEC1_neg', 
-
-                                                        'TEC1_pos+TEC2_pos', 
-                                                        'TEC2_pos+TEC3_pos', 
-                                                        'TEC3_pos+TEC4_pos',
-                                                        'TEC4_pos+TEC5_pos',
-                                                        'TEC5_pos+TEC6_pos',
-                                                        'TEC6_pos+TEC7_pos',
-                                                        'TEC7_pos+TEC8_pos',
-
-                                                        'TEC1_neg+TEC2_neg', 
-                                                        'TEC2_neg+TEC3_neg', 
-                                                        'TEC3_neg+TEC4_neg',
-                                                        'TEC4_neg+TEC5_neg',
-                                                        'TEC5_neg+TEC6_neg',
-                                                        'TEC6_neg+TEC7_neg',
-                                                        'TEC7_neg+TEC8_neg'
-                                                        #other combinations could be added
-                                                        ),
-
+    
+             
+convLayerPairs = cms.EDProducer("SeedingLayersEDProducer",
+                                layerList = cms.vstring(_convLayerPairsLayerList),
                                 BPix = cms.PSet(
                                     TTRHBuilder = cms.string('WithTrackAngle'),
                                     HitProducer = cms.string('siPixelRecHits'),
@@ -222,92 +225,33 @@ convLayerPairs = cms.EDProducer("SeedingLayersEDProducer",
 #this was done by Sam Harper (Aug 2017) and is a best guess (and I do guess wrong sometimes)
 #in general I kept all the old layer pairs and added sensible phaseI combinations
 #most pairs are adjacent layers (with some small exceptions) so I stuck to that
+_convLayerPairsLayerListPhaseI = ['BPix1+BPix2', 
+                                  
+                                  'BPix2+BPix3', 
+                                  'BPix3+BPix4',#added for PhaseI
+                                  
+                                  'BPix2+FPix1_pos', 
+                                  'BPix2+FPix1_neg', 
+                                  'BPix2+FPix2_pos', 
+                                  'BPix2+FPix2_neg',                                        
+                                  'BPix3+FPix1_pos', #added for phaseI 
+                                  'BPix3+FPix1_neg', #added for phaseI 
+                                  
+                                  'FPix1_pos+FPix2_pos', 
+                                  'FPix1_neg+FPix2_neg',
+                                  
+                                  'FPix2_pos+FPix3_pos', #added for phaseI 
+                                  'FPix2_neg+FPix3_neg',#added for phaseI 
+                                  
+                                  'BPix3+TIB1', 
+                                  #'BPix3+TIB2' #removed for phaseI
+                                  'BPix4+TIB1', #added for phase I
+                                  'BPix4+TIB2', #added for phase I
+                                  ] 
+_convLayerPairsLayerListPhaseI.extend(_convLayerPairsStripOnlyLayers)
+
 from Configuration.Eras.Modifier_trackingPhase1_cff import trackingPhase1
-trackingPhase1.toModify(convLayerPairs, layerList = cms.vstring(
-        'BPix1+BPix2', 
-        
-        'BPix2+BPix3', 
-        'BPix3+BPix4',#added for PhaseI
-        
-        'BPix2+FPix1_pos', 
-        'BPix2+FPix1_neg', 
-        'BPix2+FPix2_pos', 
-        'BPix2+FPix2_neg', 
-        'BPix2+FPix3_pos', #added for PhaseI (iffy)
-        'BPix2+FPix3_neg', #added for PhaseI (iffy)
-        
-        'BPix3+FPix1_pos', #added for phaseI (iffy)
-        'BPix3+FPix1_neg', #added for phaseI (iffy)
-        'BPix3+FPix2_pos', #added for phaseI (iffy)
-        'BPix3+FPix2_neg', #added for phaseI (iffy)
-        
-        'FPix1_pos+FPix2_pos', 
-        'FPix1_neg+FPix2_neg',
-        
-        'FPix2_pos+FPix3_pos', #added for phaseI 
-        'FPix2_neg+FPix3_neg',#added for phaseI 
-                
-        'BPix3+TIB1', 
-        'BPix3+TIB2', #maybe remove for phaseI?
-        'BPix4+TIB1', #added for phase I
-        'BPix4+TIB2', #added for phase I
-        
-        'TIB1+TID1_pos', 
-        'TIB1+TID1_neg', 
-        'TIB1+TID2_pos', 
-        'TIB1+TID2_neg',
-        'TIB1+TIB2',
-        'TIB1+TIB3',
-        
-        'TIB2+TID1_pos', 
-        'TIB2+TID1_neg', 
-        'TIB2+TID2_pos', 
-        'TIB2+TID2_neg', 
-        'TIB2+TIB3',
-        'TIB2+TIB4', 
-        
-        'TIB3+TIB4', 
-        'TIB3+TOB1', 
-        'TIB3+TID1_pos', 
-        'TIB3+TID1_neg', 
-        
-        'TIB4+TOB1',
-        'TIB4+TOB2',
-        
-        'TOB1+TOB2', 
-        'TOB1+TOB3', 
-        'TOB1+TEC1_pos', 
-        'TOB1+TEC1_neg', 
-        
-        'TOB2+TOB3',  
-        'TOB2+TOB4',
-        'TOB2+TEC1_pos', 
-        'TOB2+TEC1_neg', 
-     
-        'TID1_pos+TID2_pos', 
-        'TID2_pos+TID3_pos', 
-        'TID3_pos+TEC1_pos', 
-        
-        'TID1_neg+TID2_neg', 
-        'TID2_neg+TID3_neg', 
-        'TID3_neg+TEC1_neg', 
-        
-        'TEC1_pos+TEC2_pos', 
-        'TEC2_pos+TEC3_pos', 
-        'TEC3_pos+TEC4_pos',
-        'TEC4_pos+TEC5_pos',
-        'TEC5_pos+TEC6_pos',
-        'TEC6_pos+TEC7_pos',
-        'TEC7_pos+TEC8_pos',
-        
-        'TEC1_neg+TEC2_neg', 
-        'TEC2_neg+TEC3_neg', 
-        'TEC3_neg+TEC4_neg',
-        'TEC4_neg+TEC5_neg',
-        'TEC5_neg+TEC6_neg',
-        'TEC6_neg+TEC7_neg',
-        'TEC7_neg+TEC8_neg')
-)
+trackingPhase1.toModify(convLayerPairs, layerList = cms.vstring(_convLayerPairsLayerListPhaseI))
 
 
 trackingPhase2PU140.toReplaceWith(convLayerPairs, cms.EDProducer("SeedingLayersEDProducer",