--- a/RecoTracker/ConversionSeedGenerators/python/PhotonConversionTrajectorySeedProducerFromSingleLeg_cfi.py
+++ b/RecoTracker/ConversionSeedGenerators/python/PhotonConversionTrajectorySeedProducerFromSingleLeg_cfi.py
@@ -1,5 +1,10 @@
 import FWCore.ParameterSet.Config as cms
 
+
+from RecoTracker.TkSeedGenerator.SeedGeneratorFromRegionHitsEDProducer_cfi import seedGeneratorFromRegionHitsEDProducer 
+CommonClusterCheckPSet = seedGeneratorFromRegionHitsEDProducer.ClusterCheckPSet
+
+
 photonConvTrajSeedFromSingleLeg  = cms.EDProducer("PhotonConversionTrajectorySeedProducerFromSingleLeg",
                                                   TrackRefitter        = cms.InputTag('TrackRefitter',''),
                                                   primaryVerticesTag   = cms.InputTag("offlinePrimaryVertices"),
@@ -20,14 +25,7 @@ photonConvTrajSeedFromSingleLeg  = cms.EDProducer("PhotonConversionTrajectorySee
                                                   SeedComparitorPSet = cms.PSet(
                                                       ComponentName = cms.string('none')
                                                       ),
-                                                  ClusterCheckPSet = cms.PSet(
-                                                      PixelClusterCollectionLabel = cms.InputTag("siPixelClusters"),
-                                                      MaxNumberOfCosmicClusters = cms.uint32(400000),
-                                                      doClusterCheck = cms.bool(False),
-                                                      ClusterCollectionLabel = cms.InputTag("siStripClusters"),
-                                                      MaxNumberOfPixelClusters = cms.uint32(40000),
-                                                      cut = cms.string("strip < 400000 && pixel < 40000 && (strip < 50000 + 10*pixel) && (pixel < 5000 + 0.1*strip)"),
-                                                      ),
+                                                  ClusterCheckPSet = CommonClusterCheckPSet,
                                                   RegionFactoryPSet = cms.PSet(
                                                       RegionPSet = cms.PSet( precise = cms.bool(True),
                                                                              beamSpot = cms.InputTag("offlineBeamSpot"),