A protection was against large number of clusters was mistakenly kept for the single leg conversions as alerted here: https://github.com/cms-sw/cmssw/issues/19803

This was discussed last friday in the reco meeting and it was decided modulo timing checks to remove the the cut as it was done for other tracking cuts. Tracker will re-enable it once (if, it might not be necessary) new cuts are tuned:
https://hypernews.cern.ch/HyperNews/CMS/get/recoTracking/1710/1.html

At the same time, I noticed that BPIX4 and FPIX 3 were not in the seeding layers so I have added them to convLayerPairs (only used by photonConvTrajSeedFromSingleLeg). I took reasonable guesses to what they should be, they are marked which ones were added (and if I think it was an iffy guess)

Timing:
I benchmarked the change with step3 10224.0 MC

step3Test --conditions auto:phase1_2017_realistic --pileup_input das:/RelValMinBias_13/CMSSW_9_2_3-92X_upgrade2017_realistic_v1_earlyBS2017-v1/GEN-SIM -n 10 --era Run2_2017 --eventcontent RECOSIM,MINIAODSIM --runUnscheduled -s RAW2DIGI,L1Reco,RECO,EI,PAT --datatier GEN-SIM-RECO,MINIAODSIM --pileup AVE_35_BX_25ns --geometry DB:Extended --filein file:step2.root --fileout file:step3.root

with 4 threads running, nothing else on the machine for 200 events (from the previous step2 output). 

~/tools/timeDiffFromReport.sh defaultRecoTest.log convFixRecoTest1.log 
Couldn't parse time report using CPU and wall-clock format: trying Wall-clock only
  delta/mean delta/orJob     original                   new       module name
  ---------- ------------     --------                  ----       ------------
   +1.425743      +0.00%         0.06 ms/ev ->         0.35 ms/ev offlineBeamSpot
   -1.407648      -0.01%         1.05 ms/ev ->         0.18 ms/ev hpsPFTauDiscriminationByIsolationMVArun2v1DBoldDMwLTrawBoosted
   +1.204013      +0.02%         0.60 ms/ev ->         2.40 ms/ev dt1DCosmicRecHits
   +1.161905      +0.00%         0.02 ms/ev ->         0.08 ms/ev hpsPFTauDiscriminationByVVTightIsolationMVArun2v1DBoldDMwLTBoosted
   -1.110024      -0.00%         0.32 ms/ev ->         0.09 ms/ev hpsPFTauDiscriminationByLooseIsolationMVArun2v1DBnewDMwLTBoosted
   +1.050847      +0.00%         0.03 ms/ev ->         0.09 ms/ev hpsPFTauDiscriminationByVVTightIsolationMVArun2v1DBnewDMwLTBoosted
   -0.473373      -0.01%         1.46 ms/ev ->         0.90 ms/ev QGTagger
   +0.411724      +0.00%         0.57 ms/ev ->         0.86 ms/ev totemRPRawToDigi
   +0.393310      +0.36%        62.45 ms/ev ->        93.02 ms/ev convTrackCandidates
   -0.271556      -0.00%         1.15 ms/ev ->         0.87 ms/ev shiftedPatJetResDownNoHF
   -0.269908      -0.00%         0.68 ms/ev ->         0.52 ms/ev ak4PFPuppiResidualCorrector
   +0.258348      +0.03%         8.14 ms/ev ->        10.55 ms/ev convStepTracks
   -0.250735      -0.00%         1.15 ms/ev ->         0.89 ms/ev shiftedPatSmearedJetResDownNoHF
   +0.249325      +1.34%       396.90 ms/ev ->       509.95 ms/ev RECOSIMoutput
   -0.215368      -0.01%         4.20 ms/ev ->         3.38 ms/ev pfCombinedSecondaryVertexV2BJetTags
   -0.194437      -0.04%        19.19 ms/ev ->        15.79 ms/ev gedPhotons
   +0.174399      +0.09%        37.74 ms/ev ->        44.95 ms/ev photonConvTrajSeedFromSingleLeg
   -0.172012      -0.01%         6.71 ms/ev ->         5.64 ms/ev gedGsfElectrons
   -0.161363      -0.02%        13.01 ms/ev ->        11.06 ms/ev siStripZeroSuppression
   -0.150784      -0.04%        27.12 ms/ev ->        23.32 ms/ev packedPFCandidates
   +0.088109      +0.01%        11.54 ms/ev ->        12.61 ms/ev initialStepSeeds
   +0.083235      +0.02%        15.83 ms/ev ->        17.21 ms/ev detachedTripletStepSeeds
   +0.069082      +0.05%        64.83 ms/ev ->        69.47 ms/ev pixelLessStepSeeds
   +0.055333      +0.00%         5.71 ms/ev ->         6.04 ms/ev siPixelRecHitsPreSplitting
   +0.054070      +0.05%        79.46 ms/ev ->        83.88 ms/ev conversionTrackCandidates
   +0.051940      +0.02%        30.29 ms/ev ->        31.90 ms/ev tobTecStepSeedsPair
  ---------- ------------     --------                  ----       ------------
Job total:  8.45327 s/ev ==> 8.57115 s/ev

I also looked at the file sizes:
old:
-rw-r--r-- 1 mjf21517 cms  584972464 Aug 11 16:03 recoTestStdSeeds.root
-rw-r--r-- 1 mjf21517 cms  584925605 Aug 11 16:09 recoTestStdSeeds2.root
-rw-r--r-- 1 mjf21517 cms   12465186 Aug 11 16:09 recoTestStdSeeds2_inMINIAOD.root
-rw-r--r-- 1 mjf21517 cms  584945263 Aug 11 16:09 recoTestStdSeeds3.root
-rw-r--r-- 1 mjf21517 cms   12463389 Aug 11 16:09 recoTestStdSeeds3_inMINIAOD.root
-rw-r--r-- 1 mjf21517 cms   12477855 Aug 11 16:03 recoTestStdSeeds_inMINIAOD.root
new:
-rw-r--r-- 1 mjf21517 cms  585555291 Aug 14 10:42 convFixRecoTest1.root
-rw-r--r-- 1 mjf21517 cms   12482056 Aug 14 10:42 convFixRecoTest1_inMINIAOD.root
-rw-r--r-- 1 mjf21517 cms  585498303 Aug 14 11:03 convFixRecoTest2.root
-rw-r--r-- 1 mjf21517 cms   12477193 Aug 14 11:03 convFixRecoTest2_inMINIAOD.root
-rw-r--r-- 1 mjf21517 cms  585549646 Aug 14 12:12 convFixRecoTest3.root
-rw-r--r-- 1 mjf21517 cms   12489822 Aug 14 12:12 convFixRecoTest3_inMINIAOD.root

and dont see much change, about 0.1% for reco, less for miniAOD.

Finally I looked at data. I did this in CMSSW_9_2_8 with 92X_dataRun2_Prompt_v8
cmsDriver.py  RECO -s RAW2DIGI,L1Reco,RECO,EI --runUnscheduled --nThreads 8 --data --era Run2_2017 --scenario pp --conditions 92X_dataRun2_Prompt_v6 --eventcontent RECO --datatier RECO --no_exec

I used DoubleEG dataset of run 300497 , LS 41-42 with a <PU> = 46.  DoubleEG dataset was chosen as it should be a worse case:

Couldn't parse time report using CPU and wall-clock format: trying Wall-clock only
  delta/mean delta/orJob     original                   new       module name
  ---------- ------------     --------                  ----       ------------
   +1.762437      +0.80%         7.10 ms/ev ->       112.38 ms/ev photonConvTrajSeedFromSingleLeg
   +1.759243      +1.59%        14.36 ms/ev ->       224.30 ms/ev convTrackCandidates
   +1.710926      +0.04%         0.46 ms/ev ->         5.92 ms/ev siStripZeroSuppression
   +1.700896      +0.15%         1.76 ms/ev ->        21.78 ms/ev convStepTracks
   +1.602091      +0.04%         0.63 ms/ev ->         5.68 ms/ev generalConversionStepConversionTrackMerger
   +1.177155      +0.01%         0.39 ms/ev ->         1.50 ms/ev dt1DCosmicRecHits
   +1.162791      +0.00%         0.02 ms/ev ->         0.07 ms/ev conversionStepConversionTrackProducer
   +1.039518      +0.05%         3.03 ms/ev ->         9.58 ms/ev csc2DRecHits
   -0.911988      -0.01%         2.01 ms/ev ->         0.75 ms/ev dt1DRecHits
   +0.691301      +0.00%         0.42 ms/ev ->         0.87 ms/ev MeasurementTrackerEventPreSplitting
   +0.587387      +0.00%         0.39 ms/ev ->         0.72 ms/ev initialStepSeedLayersPreSplitting
   +0.388389      +0.04%        11.41 ms/ev ->        16.91 ms/ev initialStepSeedsPreSplitting
   +0.117980      +0.05%        48.95 ms/ev ->        55.09 ms/ev allConversions
   +0.106899      +0.05%        60.19 ms/ev ->        66.98 ms/ev ecalMultiFitUncalibRecHit
   +0.075326      +0.08%       138.01 ms/ev ->       148.82 ms/ev initialStepTracksPreSplitting
  ---------- ------------     --------                  ----       ------------
Job total:  13.2007 s/ev ==> 13.6401 s/ev