@Sam-Harper 2017-08-14
ah one second, I forgot to runTheMatrix, will do right now