@cmsbuild 2017-08-14
A new Pull Request was created by  @Sam-Harper  for master.

It involves the following packages:

RecoTracker/ConversionSeedGenerators


@perrotta, @cmsbuild, @slava77 can you please review it and eventually sign? Thanks.
@ghellwig, @makortel, @felicepantaleo, @GiacomoSguazzoni, @rovere, @VinInn, @mschrode, @gpetruc, @ebrondol, @dgulhan this is something you requested to watch as well.
@davidlange6, @slava77 you are the release manager for this.

cms-bot commands are listed <a href="http://cms-sw.github.io/cms-bot-cmssw-cmds.html">here</a>