@makortel 2017-08-14
@Sam-Harper 
I would be interested to know how much of the timing increase is caused by the cluster multiplicity check, and how much the added layers (I hope the latter would dominate).

Do you still have the `step3_inDQM.root` files available from your ttbar+PU test? We have MultiTrackValidator monitoring also the conversion tracks, so it would be interesting to see the effect (even if apparently the fake rates etc. vs. seeding layer set does not currently give intended results because of a deficiency in configuration).