@cmsbuild 2017-08-14
+1
Tested at: feb5b2f3392f56f767a1ad1efc0dd632be51d6e3
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-20153/22279/summary.html

The following merge commits were also included on top of IB + this PR after doing git cms-merge-topic: 
https://github.com/cms-sw/cmssw/commit/8ef7e73e06eaf3499e183006f0ef93df28fa0c09
You can see more details here:
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-20153/22279/git-log-recent-commits
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-20153/22279/git-merge-result