@Sam-Harper 2017-08-15
So I did a run last night with just the cluster check enabled, no new layers.

This is data

Just disabling the cluster check:

 ---------- ------------     --------                  ----       ------------
   +1.708718      +0.63%         7.10 ms/ev ->        90.35 ms/ev photonConvTrajSeedFromSingleLeg
   +1.608580      +0.89%        14.36 ms/ev ->       132.43 ms/ev convTrackCandidates
   +1.595774      +0.11%         1.76 ms/ev ->        15.66 ms/ev convStepTracks
   +1.491292      +0.03%         0.63 ms/ev ->         4.31 ms/ev generalConversionStepConversionTrackMerger

disable cluster check -> disable cluster check + new layers
  +0.515035      +0.69%       132.43 ms/ev ->       224.30 ms/ev convTrackCandidates
   +0.327038      +0.05%        15.66 ms/ev ->        21.78 ms/ev convStepTracks
   +0.275138      +0.01%         4.31 ms/ev ->         5.68 ms/ev generalConversionStepConversionTrackMerger
   +0.217380      +0.16%        90.35 ms/ev ->       112.38 ms/ev photonConvTrajSeedFromSingleLeg

So most of the time is actually coming from disabling the cut which makes sense because it was triggering very often