@perrotta 2017-08-16
@Sam-Harper : maybe you can profit and clean up RecoTracker/ConversionSeedGenerators/python/PostLS1_ConversionStep_cff.py already in this PR, if not needed (as it looks like: it may rather adds some possible confusion, since postLS1 config is already dealt with in the usual ConversionStep_cff.py)