@Sam-Harper 2017-08-16
this does look confusing. By clean up, I assume you mean delete. I'll try that and see what happens, if its been used it should crash at that point and unit tests shouldnt complete.