@perrotta 2017-08-16
Yes, I meant delete, thus cleaning up the release ;-)

If it really crashes, then somewhere there is some ambiguos config for  
Phase1 conversions that should also get fixed

Thank you, A.

Sam-Harper <notifications@github.com> ha scritto:

> this does look confusing. By clean up, I assume you mean delete.  
> I'll try that and see what happens, if its been used it should crash  
> at that point and unit tests shouldnt complete.
>
> --
> You are receiving this because you were mentioned.
> Reply to this email directly or view it on GitHub:
> https://github.com/cms-sw/cmssw/pull/20153#issuecomment-322711806