@Sam-Harper 2017-08-17
Deleted the files, the job started and then the matrix tests succeed. Sorry for the delay got distracted by HLT  oncall stuff.

Also @makortel , the inDQM file is: https://cernbox.cern.ch/index.php/s/kkWDcbDhx6FDic4

Also I've been digging on the consequences of this. Only minimal impact on E/gamma. However this impact particle flow (single leg conversions help reduce photons faking charged hadrons) and therefore impact MET.