@cmsbuild 2017-08-17
+code-checks

Logs: https://cmssdt.cern.ch/SDT/code-checks/PR-20153/174