@makortel 2017-08-17
Sam, would you have the inDQM file also for the reference (release/IB without this PR)?