@makortel 2017-08-17
@cmsbuild, please test