@cmsbuild 2017-08-17
The tests are being triggered in jenkins.