@cmsbuild 2017-08-17
+1
Tested at: dbbd3738ccfffe4ffaf4daf7b275dead594ac8c2
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-20153/22356/summary.html

The following merge commits were also included on top of IB + this PR after doing git cms-merge-topic: 
https://github.com/cms-sw/cmssw/commit/5506301d1fd3a57e635070a9a5e52c3f5e5e5a5d
You can see more details here:
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-20153/22356/git-log-recent-commits
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-20153/22356/git-merge-result