@cmsbuild 2017-08-17
Comparison is ready
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-20153/22356/summary.html

Comparison Summary:
 - No significant changes to the logs found
 - Reco comparison results: 129 differences found in the comparisons
 - DQMHistoTests: Total files compared: 26
 - DQMHistoTests: Total histograms compared: 2653169
 - DQMHistoTests: Total failures: 307
 - DQMHistoTests: Total nulls: 0
 - DQMHistoTests: Total successes: 2652673
 - DQMHistoTests: Total skipped: 189
 - DQMHistoTests: Total Missing objects: 0
 - Checked 107 log files, 14 edm output root files, 26 DQM output files