@slava77 2017-08-17
So, for this PR we have two features:
- removal of the n-cluster check
    - this is binary and is rather uncontroversial: keeping kills reco for single leg even at moderate occupancies these days
- addition of new layer combinations for seeding 
    - here, some combinations are clear like bpix3-bpix4, but some are likely dominated by fakes like bpix2-fpix3 (assumes zero efficiency in two disks which already have fairly high redundancy fraction; and it also has a fairly large lever arm). 
    - this second option needs a reasonably clear analysis of efficiency vs fake and duplicate rate. If fakes blow up, some higher level physics objects can degrade (it's not just about CPU time or memory or disk size)